fixes for image scenario mapping and finalize path

+++ tests/ci/load_images.sh
@@ -50,15 +50,14 @@ build_yardstick_image()
             cmd="sudo $(which yardstick-img-lxd-modify) $(pwd)/tools/ubuntu-server-cloudimg-modify.sh"
 
             # Build the image. Retry once if the build fails
-            $cmd || $cmd  
-            
+            $cmd || $cmd
+
             if [ ! -f "${RAW_IMAGE}" ]; then
-                echo "${RAW_IMAGE}"
                 echo "Failed building RAW image"
                 exit 1
             fi
         fi
-    elif [[ "$DEPLOY_SCENARIO" == *"dpdk"* ]]; then
+    elif [[ "$DEPLOY_SCENARIO" == *"dpdk"* and "$RUN_BENCHMARK" == "true" ]]; then
         #create special image with DPDK and PKTGEN inside it
         if [ ! -f "${QCOW_IMAGE}" ];then
             local cmd
@@ -142,7 +141,7 @@ load_yardstick_image()
             ${EXTRA_PARAMS} \
             --file ${RAW_IMAGE} \
             yardstick-image)
-    elif [[ "$DEPLOY_SCENARIO" == *"dpdk"* ]]; then
+    elif [[ "$DEPLOY_SCENARIO" == *"dpdk"* and "$RUN_BENCHMARK" == "true" ]]; then
         output=$(eval openstack ${SECURE} image create \
             --public \
             --disk-format qcow2 \
@@ -177,7 +176,7 @@ load_yardstick_image()
 
     echo "Glance image id: $GLANCE_IMAGE_ID"
     
-    if [[ "$DEPLOY_SCENARIO" == *"dpdk"* ]]; then    
+    if [[ "$DEPLOY_SCENARIO" == *"dpdk"* and "$RUN_BENCHMARK" == "true" ]]; then
         cmd2=". $(which yardstick-img-dpdk-finalize.sh)"
         echo "Finalize script: $cmd2"
         
@@ -257,7 +256,7 @@ create_nova_flavor()
         echo
         echo "========== Creating yardstick-flavor =========="
 
-        if [[ "$DEPLOY_SCENARIO" == *"dpdk"* ]]; then
+        if [[ "$DEPLOY_SCENARIO" == *"dpdk"* and "$RUN_BENCHMARK" == "true" ]]; then
             # delete flavor, if it exists, befoare creating, in order to avoid conflict
             openstack ${SECURE} flavor delete yardstick-dpdk-flavor &> /dev/null || true
             # Create the nova flavor used by test cases with DPDK inside guest
@@ -290,7 +289,7 @@ create_nova_flavor()
 
 main()
 {
-    if [[ "$DEPLOY_SCENARIO" == *"dpdk"* ]]; then
+    if [[ "$DEPLOY_SCENARIO" == *"dpdk"* and "$RUN_BENCHMARK" == "true" ]]; then
         QCOW_IMAGE="/tmp/workspace/yardstick/yardstick-dpdk-image.img"
         RAW_IMAGE="/tmp/workspace/yardstick/yardstick-dpdk-image.tar.gz"
         if [ -f /home/opnfv/images/yardstick-dpdk-image.img ];then
